Fix issue with getBlock and persistence state
This fixes an issue where the root block editor
entity controller did not correctly call onChange
based on persistence and block equality. There
are times when controlled innerblocks are used
that we need to avoid calling onChange if the
changed blocks are not related to the root
entity.

We also make sure that controlled inner blocks
do not set pending block changes when the component
first loads.